// frontend/src/state/ducks/groups/types.ts
export enum ActionTypes {
  REQUEST_GROUP = 'REQUEST_GROUP',
  SUCCESS_GROUP = 'SUCCESS_GROUP',
  ERROR_GROUP = 'ERROR_GROUP',
  REQUEST_GROUP_EVENTS = 'REQUEST_GROUP_EVENTS',
  SUCCESS_GROUP_EVENTS = 'SUCCESS_GROUP_EVENTS',
  ERROR_GROUP_EVENTS = 'ERROR_GROUP_EVENTS',
  REQUEST_NEWLY_GROUP = 'REQUEST_NEWLY_GROUP',
  SUCCESS_NEWLY_GROUP = 'SUCCESS_NEWLY_GROUP',
  ERROR_NEWLY_GROUP = 'ERROR_NEWLY_GROUP',
}